+++ WebSocket.js
@@ -1,4 +1,6 @@
 
+import { str } from './Utils.js';
+
 function empty() {}
 
 
@@ -8,19 +10,24 @@ export class websocket {
 		this._on_open = empty;
 		this._on_close = empty;
 		this._uri = null;
+		this.ws = null; 
 	}
 
 	uri(uri) { this._uri = uri; return this; }
 	on_message(on_message) { this._on_message = on_message; return this; }
 	on_open(on_open) { this._on_open = on_open; return this; }
 	on_close(on_close) { this._on_close = on_close; return this; }
+	send(data) {
+		this.ws.send(str(data));
+		return this;
+	}
 
 	connect() {
-		let ws = new WebSocket('ws://' + this._uri);
-		ws.onopen = this._on_open;
-		ws.onmessage = this._on_message;
-		ws.onclose = this._on_close;
-		return ws;
+		this.ws = new WebSocket('ws://' + this._uri);
+		this.ws.onopen = this._on_open;
+		this.ws.onmessage = this._on_message;
+		this.ws.onclose = this._on_close;
+		return this;
 	}
 }